Move bitcrush effects to crate
