+++ bitcrush/Cargo.toml
@@ -14,6 +14,7 @@ members = ["xtask"]
 crate-type = ["cdylib", "lib"]
 
 [dependencies]
+fx = { path = "../fx" }
 # Remove the `assert_process_allocs` feature to allow allocations on the audio
 # thread in debug builds.
 nih_plug = { git = "https://github.com/robbert-vdh/nih-plug.git", features = ["assert_process_allocs", "standalone"] }
